Move BackStackState -> BackStack.State
